requirements.txt
discord.py
PyNaCl
discord-ext-voice-recv
python-dotenv
faster-whisper